+++ src/Session/src/SessionInterface.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Session;
@@ -25,7 +18,7 @@ interface SessionInterface
      *
      * @throws SessionException
      */
-    public function resume();
+    public function resume(): void;
 
     /**
      * Current session ID. Null when session is destroyed.